+++ src/kard/core/client_wrapper.py
@@ -32,7 +32,7 @@ class BaseClientWrapper:
             "X-Fern-Runtime": f"python/{platform.python_version()}",
             "X-Fern-Platform": f"{platform.system().lower()}/{platform.release()}",
             "X-Fern-SDK-Name": "kard-financial-sdk",
-            "X-Fern-SDK-Version": "20.7.0",
+            "X-Fern-SDK-Version": "20.8.0",
             **(self.get_custom_headers() or {}),
         }
         token = self._get_token()