+++ src/kard/notifications/types/notification_data_union.py
@@ -10,6 +10,7 @@ from ...core.pydantic_utilities import IS_PYDANTIC_V2, UniversalBaseModel
 from ...files.types.file_metadata_attribute import FileMetadataAttribute
 from .audit_update_attributes import AuditUpdateAttributes
 from .audit_update_relationships import AuditUpdateRelationships
+from .earned_reward_notification_attributes import EarnedRewardNotificationAttributes
 from .earned_reward_rejected_attributes import EarnedRewardRejectedAttributes
 from .earned_reward_relationships import EarnedRewardRelationships
 from .earned_reward_settled_attributes import EarnedRewardSettledAttributes
@@ -20,7 +21,6 @@ from .failed_transaction_relationships import FailedTransactionRelationships
 from .push_notification_placement_file_attributes import PushNotificationPlacementFileAttributes
 from .push_notification_placement_file_relationships import PushNotificationPlacementFileRelationships
 from .rejected_transaction_relationships import RejectedTransactionRelationships
-from .reward_notification_attributes import RewardNotificationAttributes
 from .transaction_relationships import TransactionRelationships
 from .valid_transaction_attributes import ValidTransactionAttributes
 
@@ -28,7 +28,7 @@ from .valid_transaction_attributes import ValidTransactionAttributes
 class NotificationDataUnion_EarnedRewardApproved(UniversalBaseModel):
     type: typing.Literal["earnedRewardApproved"] = "earnedRewardApproved"
     id: str
-    attributes: RewardNotificationAttributes
+    attributes: EarnedRewardNotificationAttributes
     relationships: EarnedRewardRelationships
 
     if IS_PYDANTIC_V2: