+++ src/kard/notifications/types/user_reward.py
@@ -3,19 +3,20 @@
 import typing
 
 import pydantic
+from ...commons.types.commission_type import CommissionType
 from ...core.pydantic_utilities import IS_PYDANTIC_V2, UniversalBaseModel
-from .earned_reward_notification_attributes import EarnedRewardNotificationAttributes
-from .earned_reward_relationships import EarnedRewardRelationships
 
 
-class EarnedRewardApprovedData(UniversalBaseModel):
-    id: str = pydantic.Field()
+class UserReward(UniversalBaseModel):
+    type: CommissionType = pydantic.Field()
     """
-    The internal ID of the notification
+    The type of reward (% or a flat $)
     """
 
-    attributes: EarnedRewardNotificationAttributes
-    relationships: EarnedRewardRelationships
+    value: float = pydantic.Field()
+    """
+    The reward value
+    """
 
     if IS_PYDANTIC_V2:
         model_config: typing.ClassVar[pydantic.ConfigDict] = pydantic.ConfigDict(extra="allow", frozen=True)  # type: ignore # Pydantic v2